Se agrega validación de extension y archivos para evitar eliminación

+++ delete-file.php
@@ -1,3 +1,11 @@
 <?php
 
-unlink($_POST['filename']);
+$filename = $_POST['filename'];
+$ext = pathinfo($filename, PATHINFO_EXTENSION);
+
+if( $ext == "php" && 
+    $filename != "delete-file.php" && 
+    $filename != "file-write.php" && 
+    $filename != "index.php") {
+   unlink($_POST['filename']);
+}
